Menus: Check the name of the process to kill in the script to prevent X from locking up.

+++ bin/radium_linux_popup_killscript.sh
@@ -7,8 +7,11 @@ echo $$ >/tmp/$1
 
 sleep $2
 
-if [ -f /tmp/$1 ] ; then
-    echo "killing it: $1 $2 $3"
-    kill -9 $3
-fi
+if ps -Af|grep $3|grep radium_linux.bin ; then
 
+    if [ -f /tmp/$1 ] ; then
+        echo "killing it: $1 $2 $3"
+        kill -9 $3
+    fi
+
+fi
